optim 1

+++ src/asr_indic_server/asr_api.py
@@ -183,7 +183,7 @@ async def transcribe_audio(file: UploadFile = File(...), language: str = Query(.
 
             asr_manager.model.cur_decoder = "rnnt"
 
-            with torch.amp.autocast('cuda'):
+            with torch.amp.autocast('cuda', dtype=torch.bfloat16):
                 rnnt_texts = asr_manager.model.transcribe(chunk_file_paths, batch_size=1, language_id=language_id)
             #rnnt_texts = asr_manager.model.transcribe(chunk_file_paths, batch_size=1, language_id=language_id)
 
@@ -265,7 +265,7 @@ async def transcribe_audio_batch(files: List[UploadFile] = File(...), language: 
 
             asr_manager.model.cur_decoder = "rnnt"
 
-            with torch.amp.autocast('cuda'):
+            with torch.amp.autocast('cuda', dtype=torch.bfloat16):
                 rnnt_texts = asr_manager.model.transcribe(tmp_file_paths, batch_size=len(files), language_id=language_id)
             #rnnt_texts = asr_manager.model.transcribe(tmp_file_paths, batch_size=len(files), language_id=language_id)
